chore(utils): Move error assertion functions to testutils
